refactoring

+++ main.c
@@ -8,6 +8,7 @@
 #include <sys/stat.h>
 #include <sys/socket.h>
 #include <sys/wait.h>
+#include <sys/sendfile.h>
 
 #define PORT 80
 #define MAX_DOCUMENT_ROOT_NAME 256
@@ -20,16 +21,17 @@ int preforks;
 
 pid_t *child_pids;
 
+pid_t parent_pid;
 
 void perror_exit(const char *str);
 
-void send_str(int socket_fd, char *str);
+inline void send_str(int socket_fd, char *str);
 
 size_t receive_line(int socket_fd, char *buffer, size_t buffer_size);
 
 void exchange_connection(int socket_fd);
 
-int get_file_size(int fd);
+off_t get_file_size(int fd);
 
 void child_doing(int socket_fd);
 
@@ -38,10 +40,8 @@ void handle_SIGCHILD(int signal) {
 }
 
 void handle_kill(int signal) {
-    pid_t self_pid = getpid();
-    for (int i = 0; i < preforks; ++i)
-        if (child_pids[i] == self_pid)
-            return;
+    if (parent_pid != getpid())
+        return;
     for (int i = 0; i < preforks; ++i)
         kill(child_pids[i], 9);
 }
@@ -73,7 +73,7 @@ int main(int argc, char *argv[]) {
         perror_exit("Bind socket");
 
     // backlog <= /proc/sys/net/core/somaxconn
-    if (listen(socket_fd, atoi(argv[2])))
+    if (listen(socket_fd, atoi(argv[2])) == -1)
         perror_exit("Listen socket");
 
     signal(SIGCHLD, handle_SIGCHILD);
@@ -84,6 +84,7 @@ int main(int argc, char *argv[]) {
 
     preforks = atoi(argv[3]);
     child_pids = malloc(sizeof(pid_t) * preforks);
+    parent_pid = getpid();
     for (int i = 0; i < preforks; ++i) {
         pid_t pid = fork();
         if (pid == 0)
@@ -99,8 +100,8 @@ int main(int argc, char *argv[]) {
 }
 
 void child_doing(int socket_fd) {
-    int client_socket_fd;
     while (1) {
+        int client_socket_fd;
         struct sockaddr_in client_address;
         socklen_t sockaddr_in_len = sizeof(struct sockaddr_in);
         if ((client_socket_fd = accept(socket_fd, (struct sockaddr *) &client_address, &sockaddr_in_len)) == -1)
@@ -111,7 +112,8 @@ void child_doing(int socket_fd) {
 }
 
 void exchange_connection(int socket_fd) {
-    char request_buffer[MAX_BUFFER_SIZE], resource_file_pass[MAX_BUFFER_SIZE], *buffer_pointer;
+    char request_buffer[MAX_BUFFER_SIZE], resource_file_pass[MAX_BUFFER_SIZE], end_connection_buffer[MAX_BUFFER_SIZE],
+            *buffer_pointer;
     const char default_file_name[] = "index.html";
     receive_line(socket_fd, request_buffer, MAX_BUFFER_SIZE - 10 - 1);
     if ((buffer_pointer = strstr(request_buffer, " HTTP/")) == 0) {
@@ -133,26 +135,31 @@ void exchange_connection(int socket_fd) {
         strcat(buffer_pointer, default_file_name);
     strcpy(resource_file_pass, document_root);
     strcat(resource_file_pass, buffer_pointer);
+
     int file_fd = open(resource_file_pass, O_RDONLY);
-    int file_length;
+    off_t file_length;
+    char file_length_str[20];
     if (file_fd == -1 || (file_length = get_file_size(file_fd)) == -1)
         send_str(socket_fd, "HTTP/1.0 404 Not Found\r\n"
-                            "Server: "SERVER_NAME"\r\n\r\n"
+                            "Connection: close\r\n"
+                            "Content-Length: 80\r\n\r\n"
                             "<html><head><title>404 Not Found</title></head><body>404 Not Found</body></html>");
     else {
+        sprintf(file_length_str, "%ld", file_length);
         send_str(socket_fd, "HTTP/1.0 200 OK\r\n"
-                            "Server: "SERVER_NAME"\r\n\r\n");
-        if (buffer_pointer == request_buffer + 4) {
-            char *file;
-            if ((file = malloc(file_length)) != NULL) {
-                read(file_fd, file, file_length);
-                write(socket_fd, file, file_length);
-                free(file);
-            }
-        }
+                            "Server: "SERVER_NAME"\r\n"
+                            "Connection: close\r\n"
+                            "Content-Length: ");
+        send_str(socket_fd, file_length_str);
+        send_str(socket_fd, "\r\n\r\n");
+        if (buffer_pointer == request_buffer + 4)
+            sendfile(socket_fd, file_fd, 0, file_length);
     }
     close(file_fd);
-    shutdown(socket_fd, SHUT_RDWR);
+
+    shutdown(socket_fd, SHUT_WR);
+    while (read(socket_fd, &end_connection_buffer, sizeof(end_connection_buffer)) != 0);
+    close(socket_fd);
 }
 
 void perror_exit(const char *str) {
@@ -160,7 +167,7 @@ void perror_exit(const char *str) {
     exit(1);
 }
 
-__attribute__((always_inline)) void send_str(int socket_fd, char *str) {
+inline void send_str(int socket_fd, char *str) {
     write(socket_fd, str, strlen(str));
 }
 
@@ -182,7 +189,7 @@ size_t receive_line(int socket_fd, char *buffer, size_t buffer_size) {
     return 0;
 }
 
-int get_file_size(int fd) {
+off_t get_file_size(int fd) {
     struct stat st;
     if (fstat(fd, &st) == -1)
         return -1;
